Properly initialize the packetno field when returning header packets, and start counting data packet sequence numbers with 3 instead of zero so that all the ogg packets come out of the encoder with the same packetno fields they would have in coming out of an Ogg file (as read by libogg, anyway).
Previously data packets started counting from zero, and the packetno
field of the returned header packets was uninitialized.

We had two choices in resolving this. One was just to add 3 to the
vorbis_dsp_state and/or vorbis_block sequence field when writing it to
the ogg_packet packetno field. The other is to actually keep the
packetno in the internal sequence fields. I chose the later based on the
fact that this same field gets set directly from the ogg_packet packetno
on decode, so things are consistent.


svn path=/trunk/vorbis/; revision=9513