autotools: Use AC_CANONICAL_HOST
We were calling AC_CANONICAL_TARGET in configure.ac under the
mistaken impression that this allowed setting the target
architecture for cross-compilation output.

However, in GNU terminology the system type that this the
target of cross-compilation is the "host" and "target"
is reserved for setting what the output of the compiled
application should itself target, when compiling a
cross-compiler. Finally "build" is the system type the
build system is itself running under.

So to cross-compile, one invokes `./configure --host ...`
not `--target`.

In any case, we only make use of the `host` variable,
so that is the macro we need to invoke to provide it.

Thanks to Maarten Bosmans for reporting the issue.

Signed-off-by: Marvin Scholz <epirat07@gmail.com>