Update version and copyright for the 1.3.7 release.
Bump soname version and vendor string for the new release.

Update the top-level copyright dates for the release.

The have only been bug fixes to the encoder, no tuning changes,
but we updated the codename last time. so doing so again this
time. It's better to be able to distinguish releases than not.

Signed-off-by: Mark Harris <mark.hsj@gmail.com>
Signed-off-by: Thomas Daede <daede003@umn.edu>