Rebuild specification.
The tools to build the Vorbis format specification are a heavy
dependency, so we keep a built copy of that documentation in
the repository for the convenience of casual developers.

Rebuild the documents to bump the date and footer link
for the 1.3.7 release.

The spec build is broken on current texlive releases, with symptoms
like a missing table of contents in the pdf and incorrect size
attributes on figure images in the html output. This version
was build in a debian:9 (stretch) container to recover more
of the expected behaviour.

Signed-off-by: Mark Harris <mark.hsj@gmail.com>