Update link to crc_v3.txt.
Ross William's website is still up, but the section containing
the CRC introduction is no longer available. Instead link to
a copy on zlib.org which hopefully will remain available.

Also remove the mailto: link. Email is not one of the included
contact links on the author's site, so it seems rude to continue
to publish it here.

Signed-off-by: Mark Harris <mark.hsj@gmail.com>