Fix tests on MSVC.
Copy the fallback M_PI define from lib/os.h to test/util.c so
the symbol is available under Microsoft's compiler, which doesn't
supply it by default.

(This can be changed by defining _USE_MATH_DEFINES before including math.h.)

The os.h fallback is used when compiling the main codec library,
but is not used in the external test harness, so compiling that
was failing with CMake under Windows.

Signed-off-by: evpobr <evpobr@gmail.com>